Implemented local-max-amplitude reading via python
Another bit of a hack to get reading of wav files implemented. Added a
python script, read_wav.py, which reads a wav file argument, resamples
and pads it so that the data approximately matches the data that CREPE
uses, and calculates the local amplitude (absolute value) maximum for
each 10-millisecond section of the file, centered on 10-millisecond
intervals (e.g. the output for time=0.050 seconds is the maximum
amplitude for the time interval [0.045, 0.055]). The script then saves
these maximum values to a csv file. nodejs then reads that csv to get
amplitude values that it can use to determine where notes are played,
based on when the amplitude is at a local maximum. It can determine the
note played by the frequency measured at (or immediately following) that
time.