Finished getting lick amplitude & frequency data
Now everything is in a nice format. It's hacky as heck, but will work as
an MVP to allow us to work with the frequency & amplitude data so that
we can get some actual tabs generated. Additionally, it is strongly
isolated from the rest of the code:
Input: Lick (of which only the audioFileLocation is needed)
Output: {time[], frequency[], confidence[], peakAmplitude[]}

The internals of the system for getting that array can certainly be
improved, and perhaps we'll eventually want to be able to pass
parameters to modify the tabbing algorithm on the fly. But for now, it's
a black box that gets the right data approximately correctly.