Added ugly-but-working tab generator
Tabs are created and printed to console after CREPE and amplitude is
determined and local maximums are found. The code to do this is
super-ugly and pretty hardcoded (e.g. uses one horizontal character per
100 milliseconds, no choice about it), but it does work. A lot of
cleaning up needs to be done, and probably a better data format should
be worked out, as well as a faster algorithm for tabbing, but ultimately
this is really the end of the MVP for tabbing.

Storing the tab in the database is not yet complete, and at present
tabbing is run in series with the response, (as an async function, so
the server keeps going) so the frontend user has to wait for tabbing to
be complete before the HTTP response is sent.