Reworked tab data structure & onset detection
Tab data structure is more clear now - each timestamp (i.e. 10-ms
interval) either is null or has a StringFret object representing the
string that was played at that timestamp, and what fret was being held
on that string at that time.

Onset detection is still terrible, but it is marginally better than
before. It'll need a full rework, probably including reading the WAV
file in javascript directly, or else doing the entire onset detection in
python (which is also viable).